Enable yardoc syntax within puppet type (Closes: #8)
Most of the syntax definitions for yardoc are self-contained and only a
tiny amount of glue to the ruby syntax is necessary.

We can use this to our advantage to implement Yard highlighting within
files of the puppet type, too. vim-puppet's syntax highlighting file has
enough similarities with the ruby one that it's very easy to implement:
we can simply change the glue parts to tie in to the similarly named
puppet syntax matches and clusters instead.